Rename files to match the new type names
