+++ core/runtime/src/types/iterator.rs
