+++ core/runtime/src/types/native_function.rs
